fix(agent): pin explicit edit bypass in super-YOLO, move inherit helper to barrel
setDelegatedWorkApprovalBypasses skipped the tool-edit write when the
stream already reported bypassed — but that can now be ancestry-resolved
inheritance, so super-YOLO on an inheriting child never pinned its own
entry and evaporated when the parent re-gated. Write it unconditionally.
Also move inheritApprovalBypassesOnChildStream beside the other cross-kind
helpers in the approval barrel.
