+++ src/tools/approval/toolEditApproval.ts
@@ -505,31 +505,3 @@ export async function requestAndWriteApprovedEdit(request: {
   );
   return { approval, ...written };
 }
-
-/**
- * Link a freshly resolved child subagent stream to its parent for bash and
- * tool-edit bypass resolution.
- *
- * A child delegated by a parent that auto-runs bash or auto-approves edits
- * should do the same — and should keep following the parent when either
- * bypass is toggled *after* the child stream already started, since this
- * registers live ancestry links rather than copying the parent's values once
- * at child creation (see `registerStreamParent`). Ancestry is tracked per
- * bypass kind, so the CLI's distinct AUTO-BASH / AUTO-APPROVE grants are
- * respected: a parent with AUTO-BASH but edits still gated propagates only
- * bash, and fresh streams default to gated either way. Delegation-proposal
- * (super-YOLO) bypass is deliberately NOT linked — a child's own delegations
- * still prompt unless granted on the child explicitly.
- */
-export function inheritApprovalBypassesOnChildStream(
-  childStreamId: StreamTabId,
-  parentStreamId?: StreamTabId,
-  session: SessionHandle = currentSession(),
-): void {
-  if (parentStreamId) {
-    session.approvals.registerStreamParent(childStreamId, parentStreamId, [
-      'bash',
-      'toolEdit',
-    ]);
-  }
-}
